Refactored code. Added package for classic apis.
