+++ apache-http-client/src/main/java/io/refactoring/http5/client/example/classic/handler/DataObjectResponseHandler.java
@@ -1,4 +1,4 @@
-package io.refactoring.http5.client.example.handler;
+package io.refactoring.http5.client.example.classic.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
