+++ apache-http-client/src/main/java/io/refactoring/http5/client/example/classic/helper/BaseHttpRequestHelper.java
@@ -1,6 +1,6 @@
-package io.refactoring.http5.client.example.helper;
+package io.refactoring.http5.client.example.classic.helper;
 
-import io.refactoring.http5.client.example.util.JsonUtils;
+import io.refactoring.http5.client.example.classic.util.JsonUtils;
 import lombok.extern.slf4j.Slf4j;
 
 /** Base HTTP request handler. */
