+++ apache-http-client/src/main/java/io/refactoring/http5/client/example/classic/helper/UserTypeHttpRequestHelper.java
@@ -1,4 +1,4 @@
-package io.refactoring.http5.client.example.helper;
+package io.refactoring.http5.client.example.classic.helper;
 
 import io.refactoring.http5.client.example.RequestProcessingException;
 import io.refactoring.http5.client.example.model.User;
@@ -8,8 +8,8 @@ import java.net.URI;
 import java.text.MessageFormat;
 import java.util.Map;
 
-import io.refactoring.http5.client.example.handler.DataObjectResponseHandler;
-import io.refactoring.http5.client.example.util.UserRequestProcessingUtils;
+import io.refactoring.http5.client.example.classic.handler.DataObjectResponseHandler;
+import io.refactoring.http5.client.example.classic.util.UserRequestProcessingUtils;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.hc.client5.http.classic.methods.HttpGet;
